Fixed non-compiling code & some typecast warnings
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@4769 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ include/Color.h
@@ -41,23 +41,23 @@ inline Uint32 NativeColourToSDLColour(Uint32 pixel, short bpp) {
 ///////////////
 // Get the specified component from the pixel (grabbed from SDL)
 inline Uint8 GetR(Uint32 pixel, SDL_PixelFormat *fmt)  {
-	return  (((pixel & fmt->Rmask) >> fmt->Rshift) << fmt->Rloss) +
-			(((pixel & fmt->Rmask) >> fmt->Rshift) >> (8 - (fmt->Rloss << 1)));
+	return  (Uint8)((((pixel & fmt->Rmask) >> fmt->Rshift) << fmt->Rloss) +
+			(((pixel & fmt->Rmask) >> fmt->Rshift) >> (8 - (fmt->Rloss << 1))));
 }
 
 inline Uint8 GetG(Uint32 pixel, SDL_PixelFormat *fmt)  {
-	return  (((pixel & fmt->Gmask) >> fmt->Gshift) << fmt->Gloss) +
-			(((pixel & fmt->Gmask) >> fmt->Gshift) >> (8 - (fmt->Gloss << 1)));
+	return  (Uint8)((((pixel & fmt->Gmask) >> fmt->Gshift) << fmt->Gloss) +
+			(((pixel & fmt->Gmask) >> fmt->Gshift) >> (8 - (fmt->Gloss << 1))));
 }
 
 inline Uint8 GetB(Uint32 pixel, SDL_PixelFormat *fmt)  {
-	return  (((pixel & fmt->Bmask) >> fmt->Bshift) << fmt->Bloss) +
-			(((pixel & fmt->Bmask) >> fmt->Bshift) >> (8 - (fmt->Bloss << 1)));
+	return  (Uint8)((((pixel & fmt->Bmask) >> fmt->Bshift) << fmt->Bloss) +
+			(((pixel & fmt->Bmask) >> fmt->Bshift) >> (8 - (fmt->Bloss << 1))));
 }
 
 inline Uint8 GetA(Uint32 pixel, SDL_PixelFormat *fmt)  {
-	return  (((pixel & fmt->Amask) >> fmt->Ashift) << fmt->Aloss) +
-			(((pixel & fmt->Amask) >> fmt->Ashift) >> (8 - (fmt->Aloss << 1)));
+	return  (Uint8)((((pixel & fmt->Amask) >> fmt->Ashift) << fmt->Aloss) +
+			(((pixel & fmt->Amask) >> fmt->Ashift) >> (8 - (fmt->Aloss << 1))));
 }
 
 extern SDL_PixelFormat* mainPixelFormat;
@@ -119,7 +119,7 @@ struct Color {
 	bool operator != ( const Color & c ) const { return ! ( *this == c ); };
 	
 	Color operator * ( float f ) const { return Color( Uint8(CLAMP(r*f,0.0f,255.0f)), Uint8(CLAMP(g*f,0.0f,255.0f)), Uint8(CLAMP(b*f,0.0f,255.0f)), a ); };
-	Color operator + ( const Color & c ) const { return Color( CLAMP(Uint16(r)+c.r,0,255), CLAMP(Uint16(g)+c.g,0,255), CLAMP(Uint16(b)+c.b,0,255), (Uint16(a)+c.a)/2 ); };
+	Color operator + ( const Color & c ) const { return Color( (Uint8)CLAMP(Uint16(r)+c.r,0,255), (Uint8)CLAMP(Uint16(g)+c.g,0,255), (Uint8)CLAMP(Uint16(b)+c.b,0,255), (Uint8)(Uint16(a)+c.a)/2 ); };
 	bool operator<(const Color& c) const {
 		if(r != c.r) return r < c.r;
 		if(g != c.g) return g < c.g;
